Various clean up for performance.
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@5894 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ cli/src/commands/org/jnode/command/common/ExprCommand.java
@@ -24,7 +24,6 @@ import java.io.PrintWriter;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
-
 import org.jnode.shell.AbstractCommand;
 
 
@@ -319,7 +318,7 @@ public class ExprCommand extends AbstractCommand {
     
     private int asInteger(Object obj) throws ExprException {
         if (obj instanceof Integer) {
-            return ((Integer) obj).intValue();
+            return (Integer) obj;
         } else {
             try {
                 return Integer.parseInt(obj.toString());
